Day 3
Those odd creaking noises sure was terrifying, we had to work fast. The elves took care of the mechanics and I quickly assembled a program to interpret the well needed diagnostic data.  The stress of death wasn't helpful, but thankfully I think we got the data needed just in time.