Day 4
Oh lord! A giant squid attached to the submarine. Luckily I managed to distract it by challenging it with some bingo. My first idea was to win, but on second thoughts it was better to let it feel good about itself and maybe be a bit more peaceful about it all (you know crushing us and such). It seemed to have worked, glad I skipped harpooning class for codecamp back in the days.